requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
pinecone-client==3.1.0
langchain==0.1.9
langchain-community==0.0.24
PyMuPDF==1.23.22
sentence-transformers>=2.6.0
python-multipart==0.0.9
requests==2.31.0
python-docx==1.1.0
langchain-groq==0.1.0
langchain-pinecone==0.0.3
langchain-huggingface==0.0.2
docx2txt==0.8
